# refactor tech_mapping