--- a/conda.recipe/meta.yaml
+++ b/conda.recipe/meta.yaml
@@ -15,7 +15,7 @@ requirements:
     - python
     - pip
     - setuptools
-    - NREL-rex>=0.2.47
+    - NREL-rex>=0.2.50
     - NREL-PySAM==2.1.4
     - packaging
     - plotting
@@ -23,7 +23,7 @@ requirements:
   run:
     - python
     - pip
-    - NREL-rex>=0.2.47
+    - NREL-rex>=0.2.50
     - NREL-PySAM==2.1.4
     - packaging
     - plotting