@grantbuster I had to update the baseline techmap due to removing the ability to set the distance threshold. Now it's computed with a user-defined dist_margin (default 1.05).
Plotted below is the difference between the old baseline and the new techmap (black dots).  The new techmap is slightly more conservative (i.e., more -1 pixels)

![Figure_1](https://user-images.githubusercontent.com/18647211/110575569-4c0e3900-811c-11eb-93dc-b99ef3ef2aa3.png)