@grantbuster 2021-03-10
The map looks great! a bit of difference on the edges of the resource domain is fine. I think we should plot up the full techmap at some point to make sure there's no gaps though. The test case is pretty ideal without terrible skew to the resource grid.