@MRossol 2021-03-10
So the -1 issue was due to Travis' lat, lons.

Performance on the test domain:
Master:
In [4]: %%time
   ...: ind = TechMapping.run(EXCL, RES, dset=None, max_workers=2)
   ...:
   ...:
CPU times: user 166 ms, sys: 188 ms, total: 354 ms
Wall time: 4.89 s

New branch:
In [2]: %%time
   ...: ind = TechMapping.run(EXCL, RES, dset=None, max_workers=2)
   ...:
   ...:
CPU times: user 78.4 ms, sys: 68.4 ms, total: 147 ms
Wall time: 3.89 s

Not a great test case, but still faster!