@grantbuster 2021-03-10
Sweet! if it works, its faster, and its more simple then hell yes.